Move core indexing functionality to _src/indexing.py
